+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 - Support for updating `assign_public_ip` attribute in `oci_core_instance` resource
 - Support for Oracle Analytics cloud
+- Support for Oracle Integration cloud
 
 ## 3.50.0 (October 30, 2019)